Yapay zeka/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
textblob>=0.17.0
openpyxl>=3.1.0
seaborn>=0.12.0
matplotlib>=3.7.0
regex>=2023.6.0
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
groq>=0.4.0
huggingface_hub>=0.19.0